feat: 헥사고날 아키텍처 구조 적립
