+++ src/user/adapter/in/controllers/dto/login.dto.ts
@@ -1,5 +1,5 @@
 import { IsEnum, IsString, ValidateNested } from 'class-validator';
-import { SignUpPlatformType } from '../type/SignUpPlatformType';
+import { SignUpPlatformType } from '../../../../type/SignUpPlatformType';
 import { Type } from 'class-transformer';
 
 export class NormalLoginDto {
@@ -13,11 +13,13 @@ export class LoginDto {
   @IsEnum(SignUpPlatformType)
   platform: SignUpPlatformType;
 
-  @ValidateNested()
-  @Type((dto) => {
-    if (dto.object.platform === 'Normal') {
-      return NormalLoginDto;
-    }
-  })
-  data: NormalLoginDto;
+  // @ValidateNested()
+  // @Type((dto) => {
+  //   if (dto.object.platform === 'Normal') {
+  //     return NormalLoginDto;
+  //   }
+  // })
+  // data: NormalLoginDto;
+  id: string;
+  pw: string;
 }
