+++ src/user/application/ports/in/sign-up.command.ts
@@ -1,0 +1,9 @@
+import { SignUpPlatformType } from "src/user/type/SignUpPlatformType";
+
+export class SignUpCommand {
+  constructor(
+    readonly platform: SignUpPlatformType,
+    readonly id: string,
+    readonly pw: string,
+  ) {}
+}
