+++ src/user/application/ports/in/sign-up.service.interface.ts
@@ -1,0 +1,5 @@
+import { SignUpCommand } from './sign-up.command';
+
+export interface ISignUpService {
+  signUp(command: SignUpCommand);
+}
