--- src/user/interface/user.service.interface.ts
@@ -1,5 +1,0 @@
-import { LoginDto } from '../dto/login.dto';
-
-export interface IUserService {
-  login(loginDto: LoginDto);
-}
